# Conflicts:
#	src/actor.h
#	src/gl/scene/gl_portal.h
#	src/gl/scene/gl_sky.cpp
#	src/gl/scene/gl_wall.h
#	src/gl/scene/gl_walls.cpp
#	src/p_local.h
#	src/p_map.cpp
#	src/p_spec.cpp
#	src/r_data/r_interpolate.cpp
#	wadsrc/static/xlat/eternity.txt

(note: this has not been tested. It may not compile!)

+++ src/menu/joystickmenu.cpp
@@ -225,7 +225,7 @@ public:
 
 	void SetSelection(int Selection)
 	{
-		float f = fabs(SELECTED_JOYSTICK->GetAxisScale(mAxis));
+		float f = fabsf(SELECTED_JOYSTICK->GetAxisScale(mAxis));
 		if (Selection) f*=-1;
 		SELECTED_JOYSTICK->SetAxisScale(mAxis, f);
 	}
